+++ quill/include/quill/backend/SignalHandler.h
@@ -21,7 +21,6 @@
 #include <cstring>
 #include <initializer_list>
 #include <string>
-#include <vector>
 
 #if defined(_WIN32)
   #if !defined(WIN32_LEAN_AND_MEAN)
@@ -62,6 +61,7 @@ public:
   std::atomic<uint32_t> lock{0};
   std::atomic<uint32_t> backend_thread_id{0};
   std::atomic<uint32_t> signal_handler_timeout_seconds{20};
+  std::atomic<bool> should_reraise_signal{true};
 
 private:
   SignalHandlerContext() = default;
@@ -115,6 +115,7 @@ void on_signal(int32_t signal_number)
   // Get the id of this thread in the handler and make sure it is not the backend worker thread
   uint32_t const backend_thread_id = SignalHandlerContext::instance().backend_thread_id.load();
   uint32_t const current_thread_id = get_thread_id();
+  bool const should_reraise_signal = SignalHandlerContext::instance().should_reraise_signal.load();
 
   if ((backend_thread_id == 0) || (current_thread_id == backend_thread_id))
   {
@@ -123,7 +124,7 @@ void on_signal(int32_t signal_number)
     {
       std::exit(EXIT_SUCCESS);
     }
-    else
+    else if (should_reraise_signal)
     {
       // for other signals expect SIGINT and SIGTERM we re-raise
       std::signal(signal_number, SIG_DFL);
@@ -133,37 +134,39 @@ void on_signal(int32_t signal_number)
   else
   {
     // This means signal handler is running on a frontend thread, we can log and flush
-    std::vector<LoggerBase*> loggers = detail::LoggerManager::instance().get_all_loggers();
+    LoggerBase* logger_base = detail::LoggerManager::instance().get_valid_logger();
 
-    if (!loggers.empty())
+    if (logger_base)
     {
 #if defined(_WIN32)
-      std::string const signal_desc = std::to_string(signal_number);
+      int32_t const signal_desc = signal_number;
 #else
-      std::string const signal_desc = ::strsignal(signal_number);
+      char const* const signal_desc = ::strsignal(signal_number);
 #endif
 
-      auto logger = reinterpret_cast<LoggerImpl<TFrontendOptions>*>(loggers.front());
-      QUILL_SIGNAL_HANDLER_LOG(logger, quill::LogLevel::Info, "Received signal: {}", signal_desc);
+      auto logger = reinterpret_cast<LoggerImpl<TFrontendOptions>*>(logger_base);
+      QUILL_SIGNAL_HANDLER_LOG(logger, quill::LogLevel::Info, "Received signal: {} (signum: {})",
+                               signal_desc, signal_number);
+      logger->flush_log(0);
 
       if (signal_number == SIGINT || signal_number == SIGTERM)
       {
         // For SIGINT and SIGTERM, we are shutting down gracefully
         // Pass `0` to avoid calling std::this_thread::sleep_for()
-        logger->flush_log(0);
         std::exit(EXIT_SUCCESS);
       }
       else
       {
-        QUILL_SIGNAL_HANDLER_LOG(logger, quill::LogLevel::Critical,
-                                 "Terminated unexpectedly because of signal: {}", signal_desc);
+        if (should_reraise_signal)
+        {
+          QUILL_SIGNAL_HANDLER_LOG(logger, quill::LogLevel::Critical,
+                                   "Program terminated unexpectedly due to signal: {} (signum: {})",
+                                   signal_desc, signal_number);
 
-        // Pass `0` to avoid calling std::this_thread::sleep_for()
-        logger->flush_log(0);
-
-        // Reset to the default signal handler and re-raise the signal
-        std::signal(signal_number, SIG_DFL);
-        std::raise(signal_number);
+          // Reset to the default signal handler and re-raise the signal
+          std::signal(signal_number, SIG_DFL);
+          std::raise(signal_number);
+        }
       }
     }
   }
@@ -238,11 +241,12 @@ BOOL WINAPI on_console_signal(DWORD signal)
       (signal == CTRL_C_EVENT || signal == CTRL_BREAK_EVENT))
   {
     // Log the interruption and flush log messages
-    std::vector<LoggerBase*> loggers = detail::LoggerManager::instance().get_all_loggers();
-    if (!loggers.empty())
+    LoggerBase* logger_base = detail::LoggerManager::instance().get_valid_logger();
+    if (logger_base)
     {
-      auto logger = reinterpret_cast<LoggerImpl<TFrontendOptions>*>(loggers.front());
-      QUILL_SIGNAL_HANDLER_LOG(logger, quill::LogLevel::Info, "Interrupted by Ctrl+C:");
+      auto logger = reinterpret_cast<LoggerImpl<TFrontendOptions>*>(logger_base);
+      QUILL_SIGNAL_HANDLER_LOG(logger, quill::LogLevel::Info,
+                               "Program interrupted by Ctrl+C or Ctrl+Break signal");
 
       // Pass `0` to avoid calling std::this_thread::sleep_for()
       logger->flush_log(0);
@@ -264,17 +268,19 @@ LONG WINAPI on_exception(EXCEPTION_POINTERS* exception_p)
   if ((backend_thread_id != 0) && (current_thread_id != backend_thread_id))
   {
     // Log the interruption and flush log messages
-    std::vector<LoggerBase*> loggers = detail::LoggerManager::instance().get_all_loggers();
-    if (!loggers.empty())
+    LoggerBase* logger_base = detail::LoggerManager::instance().get_valid_logger();
+    if (logger_base)
     {
-      auto logger = reinterpret_cast<LoggerImpl<TFrontendOptions>*>(loggers.front());
+      auto logger = reinterpret_cast<LoggerImpl<TFrontendOptions>*>(logger_base);
 
-      QUILL_SIGNAL_HANDLER_LOG(logger, quill::LogLevel::Info, "Received exception code: {}",
-                               get_error_message(exception_p->ExceptionRecord->ExceptionCode));
+      QUILL_SIGNAL_HANDLER_LOG(logger, quill::LogLevel::Info, "Received exception: {} (Code: {})",
+                               get_error_message(exception_p->ExceptionRecord->ExceptionCode),
+                               exception_p->ExceptionRecord->ExceptionCode);
 
       QUILL_SIGNAL_HANDLER_LOG(logger, quill::LogLevel::Critical,
-                               "Terminated unexpectedly because of exception code: {}",
-                               get_error_message(exception_p->ExceptionRecord->ExceptionCode));
+                               "Program terminated unexpectedly due to exception: {} (Code: {})",
+                               get_error_message(exception_p->ExceptionRecord->ExceptionCode),
+                               exception_p->ExceptionRecord->ExceptionCode);
 
       // Pass `0` to avoid calling std::this_thread::sleep_for()
       logger->flush_log(0);